- The GPU Stokes kernel is now working with the ISMIP - HOM A010 test (but not optimized yet): 	- In FEMFlowSolution: 		- Use void pointers to create the buffers on the GPU since casting from double* to float* in single precision modes does not work. Instead allocate void pointers and fill them with float or double values accordingly. 	 	  Do the same when assigning the private kernel arguments. 		- Fix an error in using clEnqueueMapBuffer()/clEnqueueUnmapMemObject() since a mapped memory should be unmapped after the memory operation is complete. 		- Make sure to copy the solution result to the GPU at each non-linear iteration. 		- Make sure to set the newton linearization term on the GPU when needed inside the non-linearity loop. 		- Make sure to initialize to zero the matrix at each non-linear iteration. 		- Fix an error where we wrongly decided to allocate a FEMNumericIntegration instance only for a serial bulk assembly. In fact we also need it for the boundary elements assembly in a parallel mode, 		- Fix a error where the wrong term was passed to the kernel options resulting in the double precision mode calculation not working. 		- Add a few utility C functions to initialize arrays and to set the precision mode. 	- In FEMMeshUtils: 		- Fix an error where we needed to save the element number of DOFs and the element node indexes in the file mesh.colored_elements after coloring the mesh. 	- Change the test setUpISMIP_HOM_A010Test_GPU to test in double precision mode since this makes the comparison with the serial assembly easier. 	- Add an initializer for the ice_flow_gpu data structure. 	- In the kernel source file: 		- Use fixed size arrays instead of arrays which size is defined at runtime. Apparently it appears to be problematic with OpenCL. 		- Fix multiple errors: missing initializations, missing pointer deference,  missing division in the calculation of the second invariant.
- Additional correction in FEMFlowSolution:
	- Fix an error in setting the newton linearization term as it should be initialized to NO for each call of the solver (i.e., in a time dependent simulation).
	- Fix an error in the serial assembly method where the value of a variable was always undefined.

- Fix an error in FEMCore_finishAssemblyModel:solution:forceVector:sizeForceVector: where the returned value from listGetString:inArray:forVariable:info: was not taken.

- Additional minor changes.

+++ Saino/FEMTest.m
@@ -744,7 +744,7 @@ static dispatch_once_t onceToken;
             [solution.solutionInfo setObject:@YES forKey:@"color mesh"];
             [solution.solutionInfo setObject:kernelPath forKey:@"gpu kernel source file"];
             [solution.solutionInfo setObject:@"ice flow" forKey:@"gpu flow type"];
-            [solution.solutionInfo setObject:@"single" forKey:@"gpu floating-point precision"];
+            [solution.solutionInfo setObject:@"double" forKey:@"gpu floating-point precision"];
             [solution.solutionInfo setObject:@(rhoi) forKey:@"gpu ice density"];
             [solution.solutionInfo setObject:@(eta) forKey:@"gpu ice viscosity"];
             [solution.solutionInfo setObject:@(gravity) forKey:@"gpu ice gravity"];
